Fixed issue #33
Solution: Copy constructors have been added to cLocation, cSequenceFeature and cGene. A static deep_copy method has been added to cAnnotatedSequence and cReferenceSequences that creates a new pointer in the cSequenceFeatureLists; m_features, m_genes and m_repeats.

Note: This has not been implemeneted to override the default copy constructor of cAnnotatedSequence or cReferenceSequences and is therefor not implemeneted in the source yet, but is available.